Constants

Constants

Core.nothing — Constant.

nothing

The singleton instance of type Nothing, used by convention when there is no value to return (as in a C void function) or when a variable or field holds no value.

source

Base.PROGRAM_FILE — Constant.

PROGRAM_FILE

A string containing the script name passed to Julia from the command line. Note that the script name remains unchanged from within included files. Alternatively see @__FILE__.

source

Base.ARGS — Constant.

ARGS

An array of the command line arguments passed to Julia, as strings.

source

Base.C_NULL — Constant.

C_NULL

The C null pointer constant, sometimes used when calling external code.

source

Base.VERSION — Constant.

VERSION

A VersionNumber object describing which version of Julia is in use. For details see Version Number Literals.

source

Base.LOAD_PATH — Constant.

LOAD_PATH

An array of paths for using and import statements to consdier as project environments or package directories when loading code. See Code Loading.

source

Base.Sys.BINDIR — Constant.

Sys.BINDIR

A string containing the full path to the directory containing the julia executable.

source

Base.Sys.CPU_CORES — Constant.

Sys.CPU_CORES

The number of logical CPU cores available in the system.

See the Hwloc.jl package for extended information, including number of physical cores.

source

Base.Sys.WORD_SIZE — Constant.

Sys.WORD_SIZE

Standard word size on the current machine, in bits.

source

Base.Sys.KERNEL — Constant.

Sys.KERNEL

A symbol representing the name of the operating system, as returned by uname of the build configuration.

source

Base.Sys.ARCH — Constant.

Sys.ARCH

A symbol representing the architecture of the build configuration.

source

Base.Sys.MACHINE — Constant.

Sys.MACHINE

A string containing the build triple.

source

See also:

stdin

stdout

stderr

ENV

ENDIAN_BOM

Libc.MS_ASYNC

Libc.MS_INVALIDATE

Libc.MS_SYNC