Markdown2

Markdown2

Documenter.Utilities.Markdown2 — Module.

...

Provides the walk function for walking over a Markdown2.MD tree.

source

Documenter.Utilities.Markdown2.List — Type.

If .orderedstart is nothing then the list is unordered. Otherwise is specifies the first number in the list.

source

Documenter.Utilities.Markdown2.MD — Type.

MD

A type to represent Markdown documents.

source

Documenter.Utilities.Markdown2.Paragraph — Type.

source

Documenter.Utilities.Markdown2.walk — Function.

walk(f, element)

Calls f(element) on element and any of its child elements. The elements are assumed to be Markdown2 elements.

source