Cards.jl Documentation

Enums

Cards.Suit — Type

Suit (Enum)

Allowable suits in poker; ordered by strength from lowest to highest.

Fields

♠

♦

♣

♥

Cards.Rank — Type

Rank (Enum)

Card ranks in poker; ordered by strength from lowest to highest.

Fields

Two -- Nine

Jack

Queen

King

Ace

Cards.PokerType — Type

PokerType (Enum)

Enumeration describing a classification of hands for poker; ordered by strength from lowest to highest.

Fields

Singles

Pair

TwoPair

ThreeOfAKind

FullHouse

FourOfAKind

Flush

Straight

StraightFlush

Types

Cards.Card — Type

Card

DATA STRUCTURE: Representation of a Card.

Fields

suit :: Suit

rank :: Rank

Constructors

Card(::Suit, ::Rank)

Cards.Deck — Type

Deck (mutable)

DATA STRUCTURE: Representation of a Desk of Cards.

Fields

cards :: Vector{Card}

left :: UInt8 (position in deck where next to draw.)

Constructors

Deck()

Cards.SPHrep — Type

SPHrep

A lower level representation of a poker hand.

A rep containing the data: [(rank, [Suits])].

These are the "ordered Card Pairings". This is a natural way that a player would organize their hand: grouping by card ranking.

This structure will be used to place a hand in a canonical way by ordering this structure in the following way: First by pairings (rank, [suits]) will be ranked by rank.

Within the pair the suits will be ranked in natural order.

Step 0: First sort second arg.

Step 1: Use the comparison function above to do the sorting.

Fields

orderedCP :: Vector{Tuple{Rank, Vector{Suit}}}

Constructors

SPHRep(ocp::Vector{Tuple{Rank, Vector{Suit}}})

Cards.PHrep — Type

PHrep

A higher level representation of a poker hand.

Essentially a rep of the form: (PokerType, SPHrep)

The SPHrep is a poker hand in canonical order – see above. This structure will allow us to compare poker hands.

Fields

pt::PokerType – The Type of hand (Single, Pair, TwoPairs, etc.)

subr::SPHrep – The lower level representation.

Constructors

PHrep(pt::PokerType, sr::SPHrep)

Functions

Cards.pokerHandSubRep — Function

pokerHandSubRep(vs)

Takes a poker hand sorted by rank from highest to lowest and converts it to a sub-representation, SPHrep.

Currently, this has the form: [(card-rank, [suits])] Here, these paired elements are sorted from highest to lowest based on the number of duplicate cards by rank.

The "suits" array keeps track of the suits of the duplicate cards and are ordered from highest to lowest.

If a tie appears (say with a hand that has type TwoPairs) sort by the larger of the card rankings of the paired elements.

Examples

Example 1: The raw (sorted by rank) hand):

julia> pokerHandSubRep([ Card(♠, Ten), Card(♣, Ten), Card(♦, Two), Card(♣, Two), Card(♥, Two) ])

    Two   -- [♥, ♣, ♦]
    Ten   -- [♣, ♠]

Example 2: The raw (sorted by rank) hand):

julia> pokerHandSubRep([ Card(♠, Ten), Card(♣, Ten), Card(♦, Ace), Card(♥, Ace), Card(♣, Jack) ])

    Ace   -- [♥, ♦]
    Ten   -- [♣, ♠]
    Jack  -- [♣]

Arguments

vs – A rank-sorted Card vector.

Returns

::SPHrep – A SPHrep with ordering as described above.

Cards.pokerHandGroupLength — Function

pokerHandGroupLength(hsRep, n)

Get the length of the nth ordered group in a poker hand.

Example: Given that we have a poker hand sub rep, hsRep, representing a full house, This function would return 3 for the call pokerHandGroupLength(hsRep, 1) and return 2 for the call pokerHandgroupLength(hsRep, 2).

Arguments

hsRep :: SPHrep – A poker sub-hand representation.

n :: Int64 – The group to examine.

Return

::Int64 – The number of Cards in the $n^{\rm th}$ group of the sub-representation.

Cards.get_card_rank_and_top_suit — Function

get_card_rank_and_top_suit(handRep)

Gets a card ranking and its top suit for the $n^{\rm th}$ element of a poker hand representation.

That is, take the nth grouping based on the PHrep(ordered as described in the PHrep doc) and return the rank and the top suit in that grouping.

Arguments

handRep::PHrep – A poker hand representation.

n::Int64 – The index into the card groupings.

Return

(Rank, Suit) – The card ranking and top suit of the nth poker hand group.

Cards.getOrderedCardsFromRep — Function

getOrderedCardsFromRep(phr)

Get the rank/suit ordered list of cards from a poker hand representation.

Arguments

phr::PHrep – The poker hand representation of a poker hand.

Return

::Vector{Card} – A new poker hand of Cards in order.

Cards.pokerTradeInCards! — Function

pokerTradeInCards!(d, h)

Given a poker hand, h, determine now many cards to replace, 0 to 2, then return the replacement. Return a copy of the original hand if cards needed. If cards are replaced, this will mutate the deck of cards, d.

Arguments

d::Deck – A deck of cards

h::Vector{Card} – A poker hand.

Return

::Tuple{::PHrep, ::PHrep} – Old and new replacement hands.

Cards.playPoker2! — Function

playPoker2!(d)

Play a game of poker with two players.

Process:

Deal each player two 5 card hands.

Compare two hands to see who would win.

Each player can then ask for up to 2 cards.

The two players are compared again to see who wins.

Note: This function mutates the deck by dealing cards to the players.

Arguments

d :: Deck – A deck of cards.

Return

nothing

Cards.pokerHand — Function

pokerHand(v)

Takes a poker hand and converts it to a representation of the form: (hand-type, poker-hand-sub-rep) See the description of poker-hand-sub-rep in the function pokerHandSubRep.

Arguments

v – A vector of Cards.

Return

::PHrep – Essentially, a pairing of hand-type with a vector with a sub-representation, currently of the form: (rank, [suits]). The sub-representation is ordered as described in the function pokerHandSubRep. Specifically, the return has the form: PHrep(hand-type, SPHrep([(card-rank, [card-suits])]))

Cards.deal! — Function

deal!(d, n)

Deal a hand of n cards from a Deck.

Note: Will mutate the deck, d.

Arguments

A – A deck of Cards.

n – The number of Cards to deal.

Return

::Vector{Cards} – A vector of n cards.

Cards.reset! — Function

reset!(d)

Set/reset the deck to be full. That is, gather any outstanding cards from any games and place them back in the deck.

Arguments

d :: Deck –A deck of Cards.

Return

nothing

Index

Cards.Card

Cards.Deck

Cards.PHrep

Cards.PokerType

Cards.Rank

Cards.SPHrep

Cards.Suit

Cards.deal!

Cards.getOrderedCardsFromRep

Cards.get_card_rank_and_top_suit

Cards.playPoker2!

Cards.pokerHand

Cards.pokerHandGroupLength

Cards.pokerHandSubRep

Cards.pokerTradeInCards!

Cards.reset!