Wordle.jl Documentation

Functions

Wordle.create_wordle_info — Function

create_wordle_info(guess, pword)

Create an information structure of the form:

([LETTER, EXACT_MATCH_POSITION)], Dict(LETTER => (NUMBER_OF_MATCHES, MATCH_FLAG))

Here, the dictionary has the in-exact match information: LETTER : A matching letter NUMBER_OF_MATCHES : The number of matches. The latter is interpreted thusly:

If MATCH_FLAG is $0$, there are exactly NUMBER_OF_MATCHES with this letter that should occur in the puzzle word.

Else there are at least NUMBER_OF_MATCHES with this letter that should occur in the puzzle word.

Arguments

guess::String: The guess for the puzzle.

pword::String: The puzzle word.

Returns

A tuple of a vector of tuples of exact matches and a dictionary of 
non-exact match info.

Examples

julia> create_wordle_info("which", "where")
([('w', 1), ('h', 2)], Dict('h' => (0, 0), 'c' => (0, 0), 'i' => (0, 0)))

julia> create_wordle_info("teens", "where")
([('e', 3)], Dict('n' => (0, 0), 's' => (0, 0), 't' => (0, 0), 'e' => (1, 1)))

Wordle.filter_universe — Function

filter_universe(wordle_info, words)

Filter an existing universe of words based on match info.

Arguments

wordle_info : Wordle info of the form: ([(LETTER, EXACT_POSITION)], Dict( LETTER => (k, n))) The Wordle info – the same type as the return value form create_wordle_info.

words : A Vector of words.

Return

A subset of the words vector based on the filter information from wordle_info.

Examples

julia> (winfo, d) = create_wordle_info("which", "where")

([('w', 1), ('h', 2)], Dict('h' => (0, 0), 'c' => (0, 0), 'i' => (0, 0)))

julia> words = ["state", "which", "where", "child", "there", "taste"]

julia> filter_universe((winfo, d), words)

1-element Vector{String}:
 "where"

Wordle.solve_wordle — Function

solve_wordle(puzzle_word, universe_df, rec_count, sol_path, last_guess, 
                lfa[; chk_inputs, guess_strategy])

Solves a Wordle puzzle.

By default, makes guesses based on the most frequently used word in the universe passed in. However, there is an option to pass in a guessing strategy function.

ASSUMES: The universe DataFrame is sorted from highest frequency to lowest.

Arguments

puzzle_word : The puzzle word.

universe_df : A DataFrame with schema: word(words of the same length), freq(freq fraction by use)

NOTE: The universe is assumed to be sorted in reverse order by the :freq column.

rec_count : The number of calls to this function.

sol_path : Any containing the current list of guesses: [ (guess, exact_info, universe_size) ...]

last_guess : The previous guess.

lfa : The lowercase alphabet listed in frequency of use order.

Keyword Arguments

chk_inputs : If true, check the input contract.

guess_strategy : If not nothing, apply this function to pick the next guess. If nothing, pick the next guess as the most frequent word in the current universe.

ul : The lower threshold size of the filtered Wordle universe.

uu : The upper threshold size of the filtered Wordle universe.

init_guess : The starting guess to use.

Here,

exact_info has the form: [(LETTER, POSITION) ...]

universe_size is the size the word list when the guess was made.

The guess_strategy is only turned on when the filtered Wordle universe is between the thresholds: ul and uu; otherwise, the default strategy is used – the most frequently used word in the existing filtered Wordle universe is chosen.

Return

(sol_path, number-of-guesses, :SUCCESS/:FAILURE)

NOTE: A sol_path that does not include the puzzle word, means that at some point after a guess was made – along with the corresponding filtering of the universe – there was only one word left. In this case the guess count was increased by 1, but the function did not recurse.

Examples

julia> solve_wordle("taste")

(Any[("their", [('t', 1)], 3027), 
     ("taken", [('t', 1), ('a', 2)], 31), 
     ("table", [('t', 1), ('a', 2), ('e', 5)], 5), 
     ("taste", [('t', 1), ('a', 2), ('s', 3), ('t', 4), ('e', 5)], 2)], 4, :SUCCESS)

Input Contract

universe_df schema is (:word, :freq)

If words is an N-vector of the words from universe_df, then

∃ m > 0, ∀ i∈[1,N], |words[i]| = m (All the words in universe_df have the same length.)

words = words[argsort[universe_df[:freq]]] (Words are sorted from highest to lowest word usage.)

Index

Wordle.create_wordle_info

Wordle.filter_universe

Wordle.solve_wordle