Wquantiles.jl Documentation

Functions

Wquantiles.wquantile — Method

wquantile(x, w, q[; chk=true, norm_wgt=true, sort_q=true])

Finds the q weighted quantile values from the vector x.

Type Constraints

S <: Real

V <: Real

Arguments

x ::Vector{T}: Vector(n) of values from which to find quantiles.

w ::Vector{S}: Vector(n) of weights to use.

q ::Vector{V}: Vector(l) of quantile values.

Keyword Args

chk ::Bool: If true, check the input contract described below.

norm_wgt::Bool: If true, normalize the weights.

NOTE: If norm_wgt is false, it is ASSUMED that w is already normalized.

sort_q ::Bool: If true, sort the quantile vector, q.

NOTE: If sort_q is false, it ASSUMED that q is already sorted.

Input Contract

The type of x implements sortable.

|x| = |w| – Length of x matches length of weights.

∀i, w[i] >= 0 – The weights are non-negative.

Σ w[i] > 0 – The sum of the weights is positive.

∀i, q[i] <= 1 – The quantile values are in $[0,1]$.

∀i, q[i] >= 0

Return

The vector(l) of weighted quantile values from x. Letting qs be the sorted quantiles of q. The entry i is the $i^{\rm th}$ quantile (in qs) of x.

Wquantiles.Wquantile — Function

Wquantile(X, w, q[; chk=true, norm_wgt=true, sort_q=true])

Finds the q weighted quantile values from the columns of the matrix X.

Type Constraints

S <: Real

V <: Real

Arguments

X ::Matrix{T}: Matrix(n,m) of values from which to find quantiles.

w ::Vector{S}: Vector(n) of weights to use.

q ::Vector{V}: Vector(l) of quantile values.

Keyword Args

chk ::Bool: If true, check the input contract described below.

norm_wgt::Bool: If true, normalize the weights.

NOTE: If norm_wgt is false, it is ASSUMED that w is already normalized.

sort_q ::Bool: If true, sort the quantile vector, q.

NOTE: If sort_q is false, it is ASSUMED that q is already sorted.

Input Contract

The type of X implements sortable.

∀i, |X[:, i]| = |w| – Length of each column of X matches length of weights.

∀i, w[i] >= 0 – Weights are non-negative.

Σ w[i] > 0 – The sum of the weights is positive.

∀i, q[i] <= 1 – The quantiles values are in $[0,1]$.

∀i, q[i] >= 0

Return

The (l,m) matrix of weighted quantile values from X. Letting qs be the sorted quantiles of q. The entry (i,j) is the $i^{\rm th}$ quantile (in qs) from the $j^{\rm th}$ column of X.

Wquantiles.WquantileM — Function

WquantileM(X, w, q[; chk=true])

Finds the q weighted quantile values from the columns of the matrix X.

Type Constraints

S <: Real

V <: Real

Arguments

X ::Matrix{T}: Matrix(n,m) of values from which to find quantiles.

w ::Vector{S}: Vector(n) of weights to use.

q ::Vector{V}: Vector(l) of quantile values.

Keyword Args

chk::Bool : If true, check the input contract described below.

Input Contract

The type of X is sortable.

∀i, |X[:, i]| = |w| – Length of each column of X matches length of weights.

∀i, w[i] >= 0 – Weights are non-negative.

Σ w[i] > 0 – The sum of the weights is positive.

∀i, 0 <= q[i] <= 1 – The quantiles values are in $[0,1]$.

Return

The (l,m) matrix of weighted quantile values from X. Letting qs be the sorted quantiles of q. The entry (i,j) is the $i^{\rm th}$ quantile (in qs) from the $j^{\rm th}$ column of X.

Index

Wquantiles.Wquantile

Wquantiles.WquantileM

Wquantiles.wquantile