Constants

Core.nothing — Constant.

nothing

The singleton instance of type Void, used by convention when there is no value to return (as in a C void function). Can be converted to an empty Nullable value.

source

Base.PROGRAM_FILE — Constant.

PROGRAM_FILE

A string containing the script name passed to Julia from the command line. Note that the script name remains unchanged from within included files. Alternatively see @__FILE__.

source

Base.ARGS — Constant.

ARGS

An array of the command line arguments passed to Julia, as strings.

source

Base.C_NULL — Constant.

C_NULL

The C null pointer constant, sometimes used when calling external code.

source

Base.VERSION — Constant.

VERSION

A VersionNumber object describing which version of Julia is in use. For details see Version Number Literals.

source

Base.LOAD_PATH — Constant.

LOAD_PATH

An array of paths as strings or custom loader objects for the require function and using and import statements to consider when loading code. To create a custom loader type, define the type and then add appropriate methods to the Base.load_hook function with the following signature:

Base.load_hook(loader::Loader, name::String, found::Any)

The loader argument is the current value in LOAD_PATH, name is the name of the module to load, and found is the path of any previously found code to provide name. If no provider has been found earlier in LOAD_PATH then the value of found will be nothing. Custom loader functionality is experimental and may break or change in Julia 1.0.

source

Base.JULIA_HOME — Constant.

JULIA_HOME

A string containing the full path to the directory containing the julia executable.

source

Core.ANY — Constant.

ANY

Equivalent to Any for dispatch purposes, but signals the compiler to skip code generation specialization for that field.

source

Base.Sys.CPU_CORES — Constant.

Sys.CPU_CORES

The number of logical CPU cores available in the system.

See the Hwloc.jl package for extended information, including number of physical cores.

source

Base.Sys.WORD_SIZE — Constant.

Sys.WORD_SIZE

Standard word size on the current machine, in bits.

source

Base.Sys.KERNEL — Constant.

Sys.KERNEL

A symbol representing the name of the operating system, as returned by uname of the build configuration.

source

Base.Sys.ARCH — Constant.

Sys.ARCH

A symbol representing the architecture of the build configuration.

source

Base.Sys.MACHINE — Constant.

Sys.MACHINE

A string containing the build triple.

source

See also:

STDIN

STDOUT

STDERR

ENV

ENDIAN_BOM

Libc.MS_ASYNC

Libc.MS_INVALIDATE

Libc.MS_SYNC

Libdl.DL_LOAD_PATH

Libdl.RTLD_DEEPBIND

Libdl.RTLD_LOCAL

Libdl.RTLD_NOLOAD

Libdl.RTLD_LAZY

Libdl.RTLD_NOW

Libdl.RTLD_GLOBAL

Libdl.RTLD_NODELETE

Libdl.RTLD_FIRST