Unicode

Unicode

Base.Unicode.isassigned — Function.

Unicode.isassigned(c) -> Bool

Returns true if the given char or integer is an assigned Unicode code point.

Examples

julia> using Unicode

julia> isassigned(101)
true

julia> isassigned('\x01')
true

source

Base.Unicode.normalize — Function.

Unicode.normalize(s::AbstractString, normalform::Symbol)

Normalize the string s according to one of the four "normal forms" of the Unicode standard: normalform can be :NFC, :NFD, :NFKC, or :NFKD. Normal forms C (canonical composition) and D (canonical decomposition) convert different visually identical representations of the same abstract string into a single canonical form, with form C being more compact. Normal forms KC and KD additionally canonicalize "compatibility equivalents": they convert characters that are abstractly similar but visually distinct into a single canonical choice (e.g. they expand ligatures into the individual characters), with form KC being more compact.

Alternatively, finer control and additional transformations may be be obtained by calling Unicode.normalize(s; keywords...), where any number of the following boolean keywords options (which all default to false except for compose) are specified:

compose=false: do not perform canonical composition

decompose=true: do canonical decomposition instead of canonical composition (compose=true is ignored if present)

compat=true: compatibility equivalents are canonicalized

casefold=true: perform Unicode case folding, e.g. for case-insensitive string comparison

newline2lf=true, newline2ls=true, or newline2ps=true: convert various newline sequences (LF, CRLF, CR, NEL) into a linefeed (LF), line-separation (LS), or paragraph-separation (PS) character, respectively

stripmark=true: strip diacritical marks (e.g. accents)

stripignore=true: strip Unicode's "default ignorable" characters (e.g. the soft hyphen or the left-to-right marker)

stripcc=true: strip control characters; horizontal tabs and form feeds are converted to spaces; newlines are also converted to spaces unless a newline-conversion flag was specified

rejectna=true: throw an error if unassigned code points are found

stable=true: enforce Unicode Versioning Stability

For example, NFKC corresponds to the options compose=true, compat=true, stable=true.

Examples

julia> using Unicode

julia> "μ" == normalize("µ", compat=true) #LHS: Unicode U+03bc, RHS: Unicode U+00b5
true

julia> normalize("JuLiA", casefold=true)
"julia"

julia> normalize("JúLiA", stripmark=true)
"JuLiA"

source

Base.Unicode.graphemes — Function.

graphemes(s::AbstractString) -> GraphemeIterator

Returns an iterator over substrings of s that correspond to the extended graphemes in the string, as defined by Unicode UAX #29. (Roughly, these are what users would perceive as single characters, even though they may contain more than one codepoint; for example a letter combined with an accent mark is a single grapheme.)

source

Base.Unicode.uppercase — Function.

uppercase(s::AbstractString)

Return s with all characters converted to uppercase.

Examples

julia> using Unicode

julia> uppercase("Julia")
"JULIA"

source

Base.Unicode.lowercase — Function.

lowercase(s::AbstractString)

Return s with all characters converted to lowercase.

Examples

julia> using Unicode

julia> lowercase("STRINGS AND THINGS")
"strings and things"

source

Base.Unicode.titlecase — Function.

titlecase(s::AbstractString) -> String

Capitalize the first character of each word in s. See also ucfirst to capitalize only the first character in s.

Examples

julia> titlecase("the Julia programming language")
"The Julia Programming Language"

source

Base.Unicode.ucfirst — Function.

ucfirst(s::AbstractString) -> String

Return s with the first character converted to uppercase (technically "title case" for Unicode). See also titlecase to capitalize the first character of every word in s.

See also: lcfirst, uppercase, lowercase, titlecase

Examples

julia> ucfirst("python")
"Python"

source

Base.Unicode.lcfirst — Function.

lcfirst(s::AbstractString)

Return s with the first character converted to lowercase.

See also: ucfirst, uppercase, lowercase, titlecase

Examples

julia> lcfirst("Julia")
"julia"

source

Base.Unicode.textwidth — Function.

textwidth(c)

Give the number of columns needed to print a character.

Examples

julia> using Unicode

julia> textwidth('α')
1

julia> textwidth('❤')
2

source

textwidth(s::AbstractString)

Give the number of columns needed to print a string.

Examples

julia> using Unicode

julia> textwidth("March")
5

source

Base.Unicode.isalnum — Function.

isalnum(c::Char) -> Bool

Tests whether a character is alphanumeric. A character is classified as alphabetic if it belongs to the Unicode general category Letter or Number, i.e. a character whose category code begins with 'L' or 'N'.

Examples

julia> using Unicode

julia> isalnum('❤')
false

julia> isalnum('9')
true

julia> isalnum('α')
true

source

Base.Unicode.isalpha — Function.

isalpha(c::Char) -> Bool

Tests whether a character is alphabetic. A character is classified as alphabetic if it belongs to the Unicode general category Letter, i.e. a character whose category code begins with 'L'.

Examples

julia> using Unicode

julia> isalpha('❤')
false

julia> isalpha('α')
true

julia> isalpha('9')
false

source

Base.Unicode.iscntrl — Function.

iscntrl(c::Char) -> Bool

Tests whether a character is a control character. Control characters are the non-printing characters of the Latin-1 subset of Unicode.

Examples

julia> using Unicode

julia> iscntrl('\x01')
true

julia> iscntrl('a')
false

source

Base.Unicode.isdigit — Function.

isdigit(c::Char) -> Bool

Tests whether a character is a decimal digit (0-9).

Examples

julia> using Unicode

julia> isdigit('❤')
false

julia> isdigit('9')
true

julia> isdigit('α')
false

source

Base.Unicode.isgraph — Function.

isgraph(c::Char) -> Bool

Tests whether a character is printable, and not a space. Any character that would cause a printer to use ink should be classified with isgraph(c)==true.

Examples

julia> using Unicode

julia> isgraph('\x01')
false

julia> isgraph('A')
true

source

Base.Unicode.islower — Function.

islower(c::Char) -> Bool

Tests whether a character is a lowercase letter. A character is classified as lowercase if it belongs to Unicode category Ll, Letter: Lowercase.

Examples

julia> using Unicode

julia> islower('α')
true

julia> islower('Γ')
false

julia> islower('❤')
false

source

Base.Unicode.isnumeric — Function.

isnumeric(c::Char) -> Bool

Tests whether a character is numeric. A character is classified as numeric if it belongs to the Unicode general category Number, i.e. a character whose category code begins with 'N'.

Note that this broad category includes characters such as ¾ and ௰. Use isdigit to check whether a character a decimal digit between 0 and 9.

Examples

julia> using Unicode

julia> isnumeric('௰')
true

julia> isnumeric('9')
true

julia> isnumeric('α')
false

julia> isnumeric('❤')
false

source

Base.Unicode.isprint — Function.

isprint(c::Char) -> Bool

Tests whether a character is printable, including spaces, but not a control character.

Examples

julia> using Unicode

julia> isprint('\x01')
false

julia> isprint('A')
true

source

Base.Unicode.ispunct — Function.

ispunct(c::Char) -> Bool

Tests whether a character belongs to the Unicode general category Punctuation, i.e. a character whose category code begins with 'P'.

Examples

julia> using Unicode

julia> ispunct('α')
false

julia> ispunct('/')
true

julia> ispunct(';')
true

source

Base.Unicode.isspace — Function.

isspace(c::Char) -> Bool

Tests whether a character is any whitespace character. Includes ASCII characters '\t', '\n', '\v', '\f', '\r', and ' ', Latin-1 character U+0085, and characters in Unicode category Zs.

Examples

julia> using Unicode

julia> isspace('\n')
true

julia> isspace('\r')
true

julia> isspace(' ')
true

julia> isspace('\x20')
true

source

Base.Unicode.isupper — Function.

isupper(c::Char) -> Bool

Tests whether a character is an uppercase letter. A character is classified as uppercase if it belongs to Unicode category Lu, Letter: Uppercase, or Lt, Letter: Titlecase.

Examples

julia> using Unicode

julia> isupper('γ')
false

julia> isupper('Γ')
true

julia> isupper('❤')
false

source

Base.Unicode.isxdigit — Function.

isxdigit(c::Char) -> Bool

Test whether a character is a valid hexadecimal digit. Note that this does not include x (as in the standard 0x prefix).

Examples

julia> using Unicode

julia> isxdigit('a')
true

julia> isxdigit('x')
false

source