TheAlgorithms

Documentation for TheAlgorithms.

TheAlgorithms.count_nucleotides

TheAlgorithms.determinant

TheAlgorithms.dna2rna

TheAlgorithms.euler_method

TheAlgorithms.is_armstrong

TheAlgorithms.line_length

TheAlgorithms.lu_decompose

TheAlgorithms.count_nucleotides — Method

Given: A DNA string s

of length at most 1000 nt.

Return: Four integers (separated by spaces) counting the respective number of times that the symbols 'A', 'C', 'G', and 'T' occur in s

TheAlgorithms.determinant — Method

Given a non singluar matrix, calculate its determinant using LU decomposition.

L and U are lower triangular and upper triangular matrices respectively such that

A = L*U

If we want to find the determinant, then

det(A) = det(LU) = det(L)*det(U)

Determinant of triangualar matrices is the product of their diagonal entries. Hence, makes finding the determinant easy.

TheAlgorithms.dna2rna — Method

Given: A DNA string t

having length at most 1000 nt.

Return: The transcribed RNA string of t

TheAlgorithms.euler_method — Function

Calculate the solution to a differential equation using forward euler method.

TheAlgorithms.is_armstrong — Method

Program to check if a number is an Armstrong/Narcissistic number in decimal system.

Armstrong number is a number that is the sum of its own digits raised to the power of the number of digits.

Contributed By:- Ashwani Rathee

A positive integer is called an Armstrong number (of order n) if

abcd... = a^n + b^n + c^n + d^n +....

TheAlgorithms.line_length — Function

Approximates the arc length of a line segment by treating the curve as a sequence of linear lines and summing their lengths.

Arguments:

f: function that returns the arc

x_start: starting x value

xend: ending xvalue

steps: steps to take for accurace, more the steps greater the accuracy

TheAlgorithms.lu_decompose — Method

Decomposes a n x n non singular matrix into a lower triangular matrix (L) and an upper triangular matrix (U)