TheAlgorithms

Documentation for TheAlgorithms.

TheAlgorithms.AbstractBinaryTree_arr

TheAlgorithms.DisjointSet

TheAlgorithms.abs_max

TheAlgorithms.abs_min

TheAlgorithms.abs_val

TheAlgorithms.area_circle

TheAlgorithms.area_ellipse

TheAlgorithms.area_heron_triangle

TheAlgorithms.area_parallelogram

TheAlgorithms.area_rectangle

TheAlgorithms.area_rhombus

TheAlgorithms.area_square

TheAlgorithms.area_trapezium

TheAlgorithms.area_triangle

TheAlgorithms.ceil_val

TheAlgorithms.celsius_to_fahrenheit

TheAlgorithms.celsius_to_kelvin

TheAlgorithms.complete_pack!

TheAlgorithms.count_nucleotides

TheAlgorithms.determinant

TheAlgorithms.dna2rna

TheAlgorithms.euler_method

TheAlgorithms.exponential_search

TheAlgorithms.factorial_iterative

TheAlgorithms.factorial_recursive

TheAlgorithms.fahrenheit_to_celsius

TheAlgorithms.fahrenheit_to_kelvin

TheAlgorithms.fcfs

TheAlgorithms.find

TheAlgorithms.floor_val

TheAlgorithms.interpolation_search

TheAlgorithms.is_armstrong

TheAlgorithms.jump_search

TheAlgorithms.kelvin_to_celsius

TheAlgorithms.kelvin_to_fahrenheit

TheAlgorithms.line_length

TheAlgorithms.lu_decompose

TheAlgorithms.mean

TheAlgorithms.median

TheAlgorithms.mode

TheAlgorithms.perfect_cube

TheAlgorithms.perfect_number

TheAlgorithms.perfect_square

TheAlgorithms.prime_check

TheAlgorithms.prime_factors

TheAlgorithms.rotation_matrix

TheAlgorithms.sum_ap

TheAlgorithms.sum_gp

TheAlgorithms.surfarea_cube

TheAlgorithms.surfarea_sphere

TheAlgorithms.zero_one_pack!

TheAlgorithms.AbstractBinaryTree_arr — Type

array-based binary tree

TheAlgorithms.DisjointSet — Type

This can contain a maximum of length(par) parenting-relations par is an array of Int, which is the index of the parent node.

TheAlgorithms.abs_max — Method

abs_max(x)

Program to find the max absolute value in a vector

Example

abs_max([1,3,4]) # returns 4
abs_max([-3,1,2]) # returns -3
abs_max([-7,-3,6]) #returns -7

Contributed By:- Ashwani Rathee

TheAlgorithms.abs_min — Method

abs_min(num)

Program to find the min absolute value in a vector

Example

abs_min([1,3,4]) # returns 1
abs_min([-3,1,2]) # returns 1
abs_min([-7,-3,6]) #returns -3

Contributed By:- Ashwani Rathee

TheAlgorithms.abs_val — Method

abs_val(num)

Program to find the absolute value of a number

Example

abs_val(-100) # returns 100
abs_val(0) # returns 0
abs(123.1) # returns 123.1
-1000 == abs_val(-1000) #returns false
1000 == abs_val(1000) #returns true

Contributed By:- Ashwani Rathee

TheAlgorithms.area_circle — Method

area_circle(radius)

Finds area of the circle

Example

area_circle(20) # returns 1256.6370614359173
area_circle(-1) # returns DomainError

TheAlgorithms.area_ellipse — Method

area_ellipse(radius_x, radius_y)

Finds area of the ellipse

Example

area_ellipse(10, 10) # returns 314.1592653589793
area_ellipse(10, 20) # returns 628.3185307179587
area_ellipse(1, -2) # returns DomainError
area_ellipse(-1, 2) # returns DomainError
area_ellipse(-1, -2) # returns DomainError

TheAlgorithms.area_heron_triangle — Method

area_heron_triangle(side1, side2, side3)

Finds area of a triangle using heron's formula

Example

area_heron_triangle(5,12,13) # returns 30.0
area_heron_triangle(-1,-2,1) # returns DomainError
area_heron_triangle(1,-2,1)  # returns DomainError
area_heron_triangle(-1,2,1)  # returns DomainError

TheAlgorithms.area_parallelogram — Method

area_parallelogram(base, height)

Finds area of the parallelogram

Example

area_parallelogram(10,20) # returns 200
area_parallelogram(-1,-2) # returns DomainError
area_parallelogram(1,-2)  # returns DomainError
area_parallelogram(-1,2)  # returns DomainError

TheAlgorithms.area_rectangle — Method

area_rectangle(length, width)

Finds area of the rectangle

Example

area_rectangle(10,20) # returns 200
area_rectangle(-1,-2) # returns DomainError
area_rectangle(1,-2)  # returns DomainError
area_rectangle(-1,2)  # returns DomainError

TheAlgorithms.area_rhombus — Method

area_rhombus(diagonal_1, diagonal_2)

Finds area of the rhombus

Example

area_rhombus(10, 20) # returns 100.0
area_rhombus(-1,-2) # returns DomainError
area_rhombus(1,-2)  # returns DomainError
area_rhombus(-1,2)  # returns DomainError

TheAlgorithms.area_square — Method

area_square(side)

Finds area of the area_square

Example

area_square(10) # returns 100
area_square(-1) # returns DomainError

TheAlgorithms.area_trapezium — Method

area_trapezium(base1,base2,height)

Finds area of the traπzium

Example

area_trapezium(10, 20, 30)  # returns 450.0
area_trapezium(-1, -2, -3)  # returns DomainError
area_trapezium(-1, 2, 3)    # returns DomainError
area_trapezium(1, -2, 3)    # returns DomainError
area_trapezium(1, 2, -3)    # returns DomainError
area_trapezium(-1, -2, 3)   # returns DomainError
area_trapezium(1, -2, -3)   # returns DomainError
area_trapezium(-1, 2, -3)   # returns DomainError

TheAlgorithms.area_triangle — Method

area_triangle(base, height)

Finds area of the right angled triangle with base height

Example

area_triangle(10,10) # returns 50.0
area_triangle(-1,-2) # returns DomainError
area_triangle(1,-2)  # returns DomainError
area_triangle(-1,2)  # returns DomainError

TheAlgorithms.ceil_val — Method

ceil_val(x)

Finds the ceiling of x as an functionInteger

Example

ceil_val(1.3)   # 2.0
ceil_val(2.0)   # returns 2.0
ceil_val(-1.5)  #returns -1.0

Reference

https://en.wikipedia.org/wiki/Floorandceiling_functions

Contributed By:- Ashwani Rathee

TheAlgorithms.celsius_to_fahrenheit — Function

celsiustofahrenheit(celsius, ndigits::Int = 2)

Converts celsius to fahrenheit and round to 2 decimal places

Example

celsius_to_fahrenheit(273.354, 3) == 524.037  # returns true
celsius_to_fahrenheit(273.354, 0) == 524.0    # returns true
celsius_to_fahrenheit(-40.0) == -40.0         # returns true
celsius_to_fahrenheit(-20.0) == -4.0          # returns true
celsius_to_fahrenheit(0) == 32.0              # returns true
celsius_to_fahrenheit(20) == 68.0             # returns true

TheAlgorithms.celsius_to_kelvin — Function

function celsiustokelvin(celsius, ndigits::Int = 2)

Converts celsius to kelvin and round to 2 decimal places

Example

celsius_to_kelvin(273.354, 3) == 546.504    # returns true
celsius_to_kelvin(273.354, 0) == 547.0      # returns true
celsius_to_kelvin(0.0) == 273.15            # returns true
celsius_to_kelvin(20.0) == 293.15           # returns true

TheAlgorithms.complete_pack! — Method

This does complete/infinite (each item can be chosen infinite times) knapsack : pack capacity = capacity weight of each item = weights value of each item = values dp array is what the function works on It returns the ans (dp[capacity])

julia> dp=zeros(Int,30)
julia> complete_pack!(20,[1,2,9],[1,3,20],dp)
43

TheAlgorithms.count_nucleotides — Method

count_nucleotides(s::AbstractString)

Given: A DNA string s

of length at most 1000 nt.

Return: Four integers (separated by spaces) counting the respective number of times that the symbols 'A', 'C', 'G', and 'T' occur in s

TheAlgorithms.determinant — Method

determinant(mat)

Given a non singluar matrix, calculate its determinant using LU decomposition.

L and U are lower triangular and upper triangular matrices respectively such that

A = L*U

If we want to find the determinant, then

det(A) = det(LU) = det(L)*det(U)

Determinant of triangualar matrices is the product of their diagonal entries. Hence, makes finding the determinant easy.

TheAlgorithms.dna2rna — Method

dna2rna(s::AbstractString)

Given: A DNA string t

having length at most 1000 nt.

Return: The transcribed RNA string of t

TheAlgorithms.euler_method — Function

euler_method(f, x0, span, h=1.0e-2)

Calculate the solution to a differential equation using forward euler method.

TheAlgorithms.exponential_search — Method

 exponential_search(arr::AbstractArray{T,1}, x::T) where {T <: Real}

Exponential Search in 1-D array Time Complexity: O(Log n)

TheAlgorithms.factorial_iterative — Method

factorial_iterative(n)

Finds factorial of a number using Iterative method

Example

factorial_iterative(5)      # returns 120
factorial_iterative(0.1)    # returns error
factorial_iterative(-1)     # returns error

Reference

factorial of a positive integer – https://en.wikipedia.org/wiki/Factorial

Contributed By:- Ashwani Rathee

TheAlgorithms.factorial_recursive — Method

factorial_recursive(n)

Finds factorial of anumber using recursive method

Example

factorial_recursive(5)      # returns 120
factorial_recursive(0.1)    # returns error
factorial_recursive(-1)     # returns error

Reference

factorial of a positive integer – https://en.wikipedia.org/wiki/Factorial

Contributed By:- Ashwani Rathee

TheAlgorithms.fahrenheit_to_celsius — Function

fahrenheittocelsius(fahrenheit, ndigits::Int = 2)

Converts fahrenheit to celsius and round to 2 decimal places

Example

fahrenheit_to_celsius(273.354, 3) == 134.086 # returns true
fahrenheit_to_celsius(273.354, 0) == 134.0   # returns true
fahrenheit_to_celsius(0.0) == -17.78         # returns true
fahrenheit_to_celsius(20.0) == -6.67         # returns true
fahrenheit_to_celsius(40.0) == 4.44          # returns true
fahrenheit_to_celsius(60.0) == 15.56         # returns true
fahrenheit_to_celsius(80.0) == 26.67         # returns true

TheAlgorithms.fahrenheit_to_kelvin — Function

fahrenheittokelvin(fahrenheit, ndigits::Int = 2)

Converts fahrenheit to kelvin and round to 2 decimal places

Example

fahrenheit_to_kelvin(273.354, 3) == 407.236 # returns true
fahrenheit_to_kelvin(273.354, 0) == 407.0   # returns true
fahrenheit_to_kelvin(0) == 255.37           # returns true
fahrenheit_to_kelvin(20.0) == 266.48        # returns true
fahrenheit_to_kelvin(40.0) == 277.59        # returns true
fahrenheit_to_kelvin(60.0) == 288.71        # returns true
fahrenheit_to_kelvin(80.0) == 299.82        # returns true

TheAlgorithms.fcfs — Method

fcfs(n, process_id, burst_time)

Implementation of first come first served scheduling algorithm

Output

Tuple of vectors (processid, bursttime, waitingtime, turnaroundtime, avgwaitingtime, avgturnaroundtime)

Example

n = 3 # number of processes
process_id = Any[1, 2, 3] # process ids
burst_times = Any[3, 4, 5] # burst times
fcfs(n, process_id, burst_times)

Reference

https://en.wikipedia.org/wiki/Scheduling(computing)#Firstcome,firstserved

Contributed By:- Ashwani Rathee

TheAlgorithms.find — Method

Find the ancestor of node x.

TheAlgorithms.floor_val — Method

floor_val(x)

Finds the floor of x as an Integer

Example

floor_val(1.3)  # 1
floor_val(2.0)  # returns 2.0
floor_val(-1.7) # returns -2.0

Reference

https://en.wikipedia.org/wiki/Floorandceiling_functions

Contributed By:- Ashwani Rathee

TheAlgorithms.interpolation_search — Method

 interpolation_search(arr::AbstractArray{T,1}, l::T, r::T, x::T) where {T <: Real}

Interpolation Search in 1-D array Time Complexity: O(log2(log2 n))

TheAlgorithms.is_armstrong — Method

is_armstrong(x)

Program to check if a number is an Armstrong/Narcissistic number in decimal system.

Armstrong number is a number that is the sum of its own digits raised to the power of the number of digits.

Contributed By:- Ashwani Rathee

A positive integer is called an Armstrong number (of order n) if

abcd... = a^n + b^n + c^n + d^n +....

TheAlgorithms.jump_search — Method

jump_search(arr::AbstractArray{T,1}, x::T, jump::T = Int(ceil(sqrt(n)))) where {T <: Real}

Jump Search in 1-D array Time Complexity : O(√ n) Time complexity of Jump Search is between Linear Search ( ( O(n) ) and Binary Search ( O (Log n) )

TheAlgorithms.kelvin_to_celsius — Function

function kelvintocelsius(kelvin, ndigits::Int = 2)

Converts kelvin to celsius and round to 2 decimal places

Example

kelvin_to_celsius(273.354, 3) == 0.204   # returns true
kelvin_to_celsius(273.354, 0) == 0.0     # returns true
kelvin_to_celsius(273.15) == 0.0         # returns true
kelvin_to_celsius(300) == 26.85          # returns true

TheAlgorithms.kelvin_to_fahrenheit — Function

function kelvintofahrenheit(kelvin, ndigits::Int = 2)

Converts kelvin to fahrenheit and round to 2 decimal places

Example

kelvin_to_fahrenheit(273.354, 3) == 32.367  # returns true
kelvin_to_fahrenheit(273.354, 0) == 32.0    # returns true
kelvin_to_fahrenheit(273.15) == 32.0        # returns true
kelvin_to_fahrenheit(300) == 80.33          # returns true

TheAlgorithms.line_length — Function

line_length(f, x_start, x_end, steps=100)

Approximates the arc length of a line segment by treating the curve as a sequence of linear lines and summing their lengths.

Arguments:

f: function that returns the arc

x_start: starting x value

xend: ending xvalue

steps: steps to take for accurace, more the steps greater the accuracy

TheAlgorithms.lu_decompose — Method

lu_decompose(mat)

Decomposes a n x n non singular matrix into a lower triangular matrix (L) and an upper triangular matrix (U)

TheAlgorithms.mean — Method

mean(nums)

Find mean of a vector of numbers

Example

mean([3, 6, 9, 12, 15, 18, 21])      # returns 12.0
mean([5, 10, 15, 20, 25, 30, 35])    # returns 20.0
mean([1, 2, 3, 4, 5, 6, 7, 8])       # returns 4.5

Contributed By:- Ashwani Rathee

TheAlgorithms.median — Method

median(nums)

Finds median of a vector of numbers

Example

median([2,1,3,4])                   # returns 2.5
median([2, 70, 6, 50, 20, 8, 4])    # returns 8
median([0])                         # returns 0

Contributed By:- Ashwani Rathee

TheAlgorithms.mode — Method

mode(nums)

Finds mode of a vector of numbers

Example

mode([2, 3, 4, 5, 3, 4, 2, 5, 2, 2, 4, 2, 2, 2])        # returns [2]
mode([3, 4, 5, 3, 4, 2, 5, 2, 2, 4, 4, 2, 2, 2])        # returns [2]
mode([3, 4, 5, 3, 4, 2, 5, 2, 2, 4, 4, 4, 2, 2, 4, 2])  # returns [2, 4]
mode(["x", "y", "y", "z"])                              # returns ["y"]
mode(["x", "x" , "y", "y", "z"])                       # returns ["x", "y"]

Contributed By:- Ashwani Rathee

TheAlgorithms.perfect_cube — Method

perfect_cube(number)

Check if a number is a perfect cube or not.

Example

perfect_cube(27) # returns true
perfect_cube(4)  # returns false

Contributed By:- Ashwani Rathee

TheAlgorithms.perfect_number — Method

perfect_number(number)

Checks if a number is a perfect_number number or not

Details

perfect_number number is a positive integer that is equal to the sum of its positive divisors, excluding the number itself.

For example : 6 is perfect_number number

Divisors of 6 => [1,2,3]

Sum of divisors => 1+2+3 = 6

6 == sum(divisors) # which is true

Example

perfect_number(27)     # returns false
perfect_number(28)     # returns true
perfect_number(496)    # returns true
perfect_number(8128)   # returns true
perfect_number(123)    # returns false

Contributed By:- Ashwani Rathee

TheAlgorithms.perfect_square — Method

perfect_square(number)

Check if a number is a perfect square or not.

Example

perfect_square(9)   # returns True
perfect_square(16)  # returns True
perfect_square(1)   # returns True
perfect_square(0)   # returns True
perfect_square(10)  # returns False

Contributed By:- Ashwani Rathee

TheAlgorithms.prime_check — Method

prime_check(number)

Checks to see if a number is a prime or not

A number is prime if it has exactly two factors: 1 and itself.

Example

prime_check(2) # returns true
prime_check(3) # returns true
prime_check(5) # returns true
prime_check(7) # returns true
prime_check(11) # returns true
prime_check(13) # returns true
prime_check(17) # returns true
prime_check(19) # returns true
prime_check(23) # returns true
prime_check(29) # returns true

Contributed By:- Ashwani Rathee

TheAlgorithms.prime_factors — Method

prime_factors(number)

Returns prime factors of number as a vector

Example

prime_factors(50)          # returns [2,5,5]
prime_factors(0)           # returns []
prime_factors(100)         # returns [2, 2, 5, 5]
prime_factors(2560)        # returns [2, 2, 2, 2, 2, 2, 2, 2, 2, 5]
prime_factors(10^-2)       # returns []
prime_factors(0.02)        # returns []
prime_factors(10^-354)     # returns []
prime_factors("hello")     # returns error

Contributed By:- Ashwani Rathee

TheAlgorithms.rotation_matrix — Method

A 2D Rotation matrix is a mtrix that rotates a vector in a 2D real space by an angle theta. For more info: https://en.wikipedia.org/wiki/Rotation_matrix

This function takes the angle theta in radians as input and returns a 2D Matrix which will rotate the the vector by angle theta.

TheAlgorithms.sum_ap — Method

sum_ap(first_term, diff, num_terms)

Finds sum of a arithmetic progression series

Input parameters

first_term : first term of the series

diff : common difference between consecutive terms

num_terms : number of terms in the series till which we count sum

Example

sum_ap(1, 1, 10)    # returns 55.0 
sum_ap(1, 10, 100)  # returns 49600.0

Contributed By:- Ashwani Rathee

TheAlgorithms.sum_gp — Method

sum_gp(first_term, ratio, num_terms)

Finds sum of n terms in a geometric progression

Input parameters

first_term : first term of the series

raio : common ratio between consecutive terms -> a2/a1 or a3/a2 or a4/a3

num_terms : number of terms in the series till which we count sum

Example

sum_gp(1, 2, 10)    # 1023.0
sum_gp(1, 10, 5)    # 11111.0
sum_gp(0, 2, 10)    # 0.0
sum_gp(1, 0, 10)    # 1.0
sum_gp(1, 2, 0)     # -0.0
sum_gp(-1, 2, 10)   # -1023.0
sum_gp(1, -2, 10)   # -341.0

Contributed By:- Ashwani Rathee

TheAlgorithms.surfarea_cube — Method

surfarea_cube(side)

Finds surface area of a cube

Example

surfarea_cube(1)  # returns 6
surfarea_cube(3)  # returns 54
surfarea_cube(-1) # returns DomainError

TheAlgorithms.surfarea_sphere — Method

surfarea_sphere(side)

Finds surface area of a sphere

Example

surfarea_sphere(5)  # returns 314.1592653589793
surfarea_sphere(1)  # returns 12.566370614359172
surfarea_sphere(-1) # returns DomainError

TheAlgorithms.zero_one_pack! — Method

zero_one_pack!(capacity::N, weights::V, values::V, dp::V) where {N <: Number,V <: AbstractVector}

This does 0-1 (each item can be chosen only once) knapsack : pack capacity = capacity weight of each item = weights value of each item = values dp array is what the function works on It returns the ans (dp[capacity])

julia> dp=zeros(Int,30)
julia> zero_one_pack!(20,[1,3,11],[2,5,30],dp)
37