TheAlgorithms

Documentation for TheAlgorithms.

TheAlgorithms.CompletePack!

TheAlgorithms.ZeroOnePack!

TheAlgorithms.abs_max

TheAlgorithms.abs_min

TheAlgorithms.abs_val

TheAlgorithms.area_circle

TheAlgorithms.area_ellipse

TheAlgorithms.area_heron_triangle

TheAlgorithms.area_parallelogram

TheAlgorithms.area_rectangle

TheAlgorithms.area_rhombus

TheAlgorithms.area_square

TheAlgorithms.area_trapezium

TheAlgorithms.area_triangle

TheAlgorithms.ceil_val

TheAlgorithms.count_nucleotides

TheAlgorithms.determinant

TheAlgorithms.dna2rna

TheAlgorithms.euler_method

TheAlgorithms.fcfs

TheAlgorithms.floor_val

TheAlgorithms.is_armstrong

TheAlgorithms.line_length

TheAlgorithms.lu_decompose

TheAlgorithms.mean

TheAlgorithms.median

TheAlgorithms.mode

TheAlgorithms.surfarea_cube

TheAlgorithms.surfarea_sphere

TheAlgorithms.CompletePack! — Method

This does complete/infinite (each item can be chosen infinite times) knapsack : pack capacity = capacity weight of each item = weights value of each item = values dp array is what the function works on It returns the ans (dp[capacity])

julia> dp=zeros(Int,30)
julia> CompletePack!(20,[1,2,9],[1,3,20],dp)
43

TheAlgorithms.ZeroOnePack! — Method

This does 0-1 (each item can be chosen only once) knapsack : pack capacity = capacity weight of each item = weights value of each item = values dp array is what the function works on It returns the ans (dp[capacity])

julia> dp=zeros(Int,30)
julia> ZeroOnePack!(20,[1,3,11],[2,5,30],dp)
37

TheAlgorithms.abs_max — Method

abs_max(x)

Program to find the max absolute value in a vector

Example

abs_max([1,3,4]) # returns 4
abs_max([-3,1,2]) # returns -3
abs_max([-7,-3,6]) #returns -7

Contributed By:- Ashwani Rathee

TheAlgorithms.abs_min — Method

abs_min(num)

Program to find the min absolute value in a vector

Example

abs_min([1,3,4]) # returns 1
abs_min([-3,1,2]) # returns 1
abs_min([-7,-3,6]) #returns -3

Contributed By:- Ashwani Rathee

TheAlgorithms.abs_val — Method

abs_val(num)

Program to find the absolute value of a number

Example

abs_val(-100) # returns 100
abs_val(0) # returns 0
abs(123.1) # returns 123.1
-1000 == abs_val(-1000) #returns false
1000 == abs_val(1000) #returns true

Contributed By:- Ashwani Rathee

TheAlgorithms.area_circle — Method

area_circle(radius)

Finds area of the circle

Example

area_circle(20) # returns 1256.6370614359173
area_circle(-1) # returns DomainError

TheAlgorithms.area_ellipse — Method

area_ellipse(radius_x, radius_y)

Finds area of the ellipse

Example

area_ellipse(10, 10) # returns 314.1592653589793
area_ellipse(10, 20) # returns 628.3185307179587
area_ellipse(1, -2) # returns DomainError
area_ellipse(-1, 2) # returns DomainError
area_ellipse(-1, -2) # returns DomainError

TheAlgorithms.area_heron_triangle — Method

area_heron_triangle(side1, side2, side3)

Finds area of a triangle using heron's formula

Example

area_heron_triangle(5,12,13) # returns 30.0
area_heron_triangle(-1,-2,1) # returns DomainError
area_heron_triangle(1,-2,1)  # returns DomainError
area_heron_triangle(-1,2,1)  # returns DomainError

TheAlgorithms.area_parallelogram — Method

area_parallelogram(base, height)

Finds area of the parallelogram

Example

area_parallelogram(10,20) # returns 200
area_parallelogram(-1,-2) # returns DomainError
area_parallelogram(1,-2)  # returns DomainError
area_parallelogram(-1,2)  # returns DomainError

TheAlgorithms.area_rectangle — Method

area_rectangle(length, width)

Finds area of the rectangle

Example

area_rectangle(10,20) # returns 200
area_rectangle(-1,-2) # returns DomainError
area_rectangle(1,-2)  # returns DomainError
area_rectangle(-1,2)  # returns DomainError

TheAlgorithms.area_rhombus — Method

area_rhombus(diagonal_1, diagonal_2)

Finds area of the rhombus

Example

area_rhombus(10, 20) # returns 100.0
area_rhombus(-1,-2) # returns DomainError
area_rhombus(1,-2)  # returns DomainError
area_rhombus(-1,2)  # returns DomainError

TheAlgorithms.area_square — Method

area_square(side)

Finds area of the area_square

Example

area_square(10) # returns 100
area_square(-1) # returns DomainError

TheAlgorithms.area_trapezium — Method

area_trapezium(base1,base2,height)

Finds area of the trapizium

Example

area_trapezium(10, 20, 30)  # returns 450.0
area_trapezium(-1, -2, -3)  # returns DomainError
area_trapezium(-1, 2, 3)    # returns DomainError
area_trapezium(1, -2, 3)    # returns DomainError
area_trapezium(1, 2, -3)    # returns DomainError
area_trapezium(-1, -2, 3)   # returns DomainError
area_trapezium(1, -2, -3)   # returns DomainError
area_trapezium(-1, 2, -3)   # returns DomainError

TheAlgorithms.area_triangle — Method

area_triangle(base, height)

Finds area of the right angled triangle with base height

Example

area_triangle(10,10) # returns 50.0
area_triangle(-1,-2) # returns DomainError
area_triangle(1,-2)  # returns DomainError
area_triangle(-1,2)  # returns DomainError

TheAlgorithms.ceil_val — Method

ceil_val(x)

Finds the ceiling of x as an Integer

Example

julia ceil_val(1.3) # 2.0 ceil_val(2.0) # returns 2.0 ceil_val(-1.5) #returns -1.0`

Reference

https://en.wikipedia.org/wiki/Floorandceiling_functions

Contributed By:- Ashwani Rathee

TheAlgorithms.count_nucleotides — Method

Given: A DNA string s

of length at most 1000 nt.

Return: Four integers (separated by spaces) counting the respective number of times that the symbols 'A', 'C', 'G', and 'T' occur in s

TheAlgorithms.determinant — Method

Given a non singluar matrix, calculate its determinant using LU decomposition.

L and U are lower triangular and upper triangular matrices respectively such that

A = L*U

If we want to find the determinant, then

det(A) = det(LU) = det(L)*det(U)

Determinant of triangualar matrices is the product of their diagonal entries. Hence, makes finding the determinant easy.

TheAlgorithms.dna2rna — Method

Given: A DNA string t

having length at most 1000 nt.

Return: The transcribed RNA string of t

TheAlgorithms.euler_method — Function

Calculate the solution to a differential equation using forward euler method.

TheAlgorithms.fcfs — Method

fcfs()

Implementation of first come first served scheduling algorithm

Output

Tuple of vectors (processid, bursttime, waitingtime, turnaroundtime, avgwaitingtime, avgturnaroundtime)

Example

n = 3 # number of processes
process_id = Any[1, 2, 3] # process ids
burst_times = Any[3, 4, 5] # burst times
fcfs(n, process_id, burst_times)

Reference

https://en.wikipedia.org/wiki/Scheduling(computing)#Firstcome,firstserved

Contributed By:- Ashwani Rathee

TheAlgorithms.floor_val — Method

floor_val(x)

Finds the floor of x as an Integer

Example

julia floor_val(1.3) # 1 floor_val(2.0) # returns 2.0 floor_val(-1.7) # returns -2.0`

Reference

https://en.wikipedia.org/wiki/Floorandceiling_functions

Contributed By:- Ashwani Rathee

TheAlgorithms.is_armstrong — Method

Program to check if a number is an Armstrong/Narcissistic number in decimal system.

Armstrong number is a number that is the sum of its own digits raised to the power of the number of digits.

Contributed By:- Ashwani Rathee

A positive integer is called an Armstrong number (of order n) if

abcd... = a^n + b^n + c^n + d^n +....

TheAlgorithms.line_length — Function

Approximates the arc length of a line segment by treating the curve as a sequence of linear lines and summing their lengths.

Arguments:

f: function that returns the arc

x_start: starting x value

xend: ending xvalue

steps: steps to take for accurace, more the steps greater the accuracy

TheAlgorithms.lu_decompose — Method

Decomposes a n x n non singular matrix into a lower triangular matrix (L) and an upper triangular matrix (U)

TheAlgorithms.mean — Method

mean(nums)

Find mean of a vector of numbers

Example

mean([3, 6, 9, 12, 15, 18, 21])      # returns 12.0
mean([5, 10, 15, 20, 25, 30, 35])    # returns 20.0
mean([1, 2, 3, 4, 5, 6, 7, 8])       # returns 4.5

Contributed By:- Ashwani Rathee

TheAlgorithms.median — Method

median(nums)

Finds median of a vector of numbers

Example

median([2,1,3,4])                   # returns 2.5
median([2, 70, 6, 50, 20, 8, 4])    # returns 8
median([0])                         # returns 0

Contributed By:- Ashwani Rathee

TheAlgorithms.mode — Method

mode(nums)

Finds mode of a vector of numbers

Example

mode([2, 3, 4, 5, 3, 4, 2, 5, 2, 2, 4, 2, 2, 2])        # returns [2]
mode([3, 4, 5, 3, 4, 2, 5, 2, 2, 4, 4, 2, 2, 2])        # returns [2]
mode([3, 4, 5, 3, 4, 2, 5, 2, 2, 4, 4, 4, 2, 2, 4, 2])  # returns [2, 4]
mode(["x", "y", "y", "z"])                              # returns ["y"]
mode(["x", "x" , "y", "y", "z"])                       # returns ["x", "y"]

Contributed By:- Ashwani Rathee

TheAlgorithms.surfarea_cube — Method

surfarea_cube(side)

Finds surface area of a cube

Example

surfarea_cube(1)  # returns 6
surfarea_cube(3)  # returns 54
surfarea_cube(-1) # returns DomainError

TheAlgorithms.surfarea_sphere — Method

surfarea_sphere(side)

Finds surface area of a sphere

Example

surfarea_sphere(5)  # returns 314.1592653589793
surfarea_sphere(1)  # returns 12.566370614359172
surfarea_sphere(-1) # returns DomainError